Update PokemonPreProcess.R
I added BoxCox trans just to see if any of the skewness was minimized, and it brought down the skewness for gymDistance quite a bit. There was an error on the box cox trans for the pokestopDistanceKm, so if you understand the error, feel free to fix it. I also did the box plots for the original numeric data, spatial sign transformation, and then the boxplots for after the spatial sign. Let me know if I did anything wrong! I could try to scale and center the data before doing spatial sign, so if you want me to do that, let me know! The only other transformation we could do could be PCA.